+++ babel.config.js
@@ -11,6 +11,7 @@ module.exports = {
     ['@babel/plugin-proposal-decorators', { legacy: true }],
     '@babel/plugin-proposal-class-properties',
     '@babel/plugin-proposal-object-rest-spread',
+    '@babel/plugin-syntax-dynamic-import',
   ],
   env: {
     test: {
